app/requirements.txt
keras
streamlit
Pillow
numpy